Block rebuild when NUMA topology changed
If the image change during a rebuild it's possible for the request
NUMA topology to change. As a rebuild uses a noop claim in the
resource tracker the NUMA topology will not be updated as part of
a rebuild.

If the NUMA constraints do not change, a rebuild will continue as normal.
If the new constraints conflict with the existing NUMA constraints of the
instance the rebuild will be rejected without altering the status of the
instance.

This change introduces an API check to block rebuild when the NUMA
requirements for the new image do not match the existing NUMA constraints.
This is in line with the previous check introduced to prevent the rebuild of
volume-backed instances which similarly are not supported.

This change adds functional tests to assert the expected behaviour of
rebuilding NUMA instances with new images. This change also asserts that
in place rebuilds of numa instances is currently not supported.

Modifications:
	nova/tests/functional/libvirt/test_numa_servers.py

NOTE(stephenfin): The new tests added in 'test_numa_servers.py' had to
be modified to use the old-style '_wait_for_state_change' function,
since change I80cdc0a33ec27b1389130c22f9c3a8ff69f6b1a0 isn't present on
'stable/train' and it's too large and invasive to justify backporting.
In addition, a 'super()' call had to be updated to use the Python 2
compatible 'super(ClassName, self)' style.

Closes-Bug: #1763766
Partial-implements: blueprint inplace-rebuild-of-numa-instances
Change-Id: I0322d872bdff68936033a6f5a54e8296a6fb3434
(cherry picked from commit 6f5358ac1992b17b7f3f99d9a32290e0d4740dae)