Avoid using image with kernel in BDM large request func test
The test_boot_from_volume_10_servers_255_volumes_2_images uses two
random images to build servers with a lot of BDMs and asserts that an
image cache is used to avoid hitting glance for the same image many
times. The functional tests uses _FakeImageService that predefines a
list of images. The test selects the first two returned by the
api.get_images() call. As the fixture stores the images in a dict the
returned image list is random in python2.7. Therefore the test runs with
two random images. One of the images
c905cedb-7281-47e4-8a62-f26bc5fc4c77 is defined to use a separate kernel
image. If this image is selected for the test then an extra glance call
is made from _handle_kernel_and_ramdisk(). This makes the test unstable
as sometimes there is 4 calls instead of the asserted (and more common)
3 calls nova.image.api.API.get().

This patch hardcodes the image_ids used in the test to avoid the
instability.

Change-Id: Ic0a9c3ffb4e52430deb26bc6ad7758105f02968b
Closes-Bug: #1847766