Add regression test for bug 1824435
This adds a regression test in our post test hook. We are not able to
do a similar test in the unit or functional tests because SQLite does
not provide any isolation between transactions on the same database
connection [1] and the bug can only be reproduced with the isolation
that is present when using a real MySQL database.

Related-Bug: #1824435

[1] https://www.sqlite.org/isolation.html

 Conflicts:
	gate/post_test_hook.sh

NOTE(melwitt): The conflict is due to change
I993c9661c37da012cc975ee8c04daa0eb9216744 which is not in Train.

Change-Id: I204361d6ff7c2323bc744878d8a9fa2d20a480b1
(cherry picked from commit 7c41365f193fd5b08d0174b6cd8e349ac95b7907)