Set instance CPU policy to 'share' through image property
Before the introduce of 'CPU resource', a 'cpu_policy' of None
is same as a 'cpu_policy' of 'shared' effectively.

But these two CPU policies ('shared' and None) have some difference now,
checking the code in 'hardware.numa_get_constraints', if the variable
'cpu_policy' is not set, which means 'cpu_policy==None', it still possible
to get the instance CPU policy through the setting of 'resources:VCPU' and
'resources:PCPU'.

We should return the 'cpu_policy==shared' if 'hw_cpu_policy==share' and
'hw:cpu_policy' is not defined. Otherwise, if you want to set a 'shared'
policy through 'hw_cpu_policy', but it might be alerted by the setting
of 'resources:PCPU'.

Change-Id: I0e2d90d83180e91c3217f5a806a48e518ab235e8
Closes-Bug: #1848308