[scheduler] Add BigVmFlavorHostSizeFilter
Big VMs are supposed to be NUMA aligned. When we don't use traits for
Placement to achieve this - which should only be the case for very old
flavors by now - we have an `extra_specs` attribute in the flavor that
specifies which fraction of a host the flavor can take. We support
providing a list of fractions and if no fraction multiplied by the
host's memory aligns with the flavor's memory, the host is not a valid
target.

We keep a list of valid fractions around in the settings.

We have to convert the values from the config to floats explicitly.
Otherwise, we only get strings.

Configuring the host flavors will include working with thirds or other
infinite fractions. Writing "0.33" is however not the same as 1/3, and
a flavor configured with a "0.33" host_fraction will not fit a
host it is arguably intended to match, because we tolerate -10%,
but +0%. Instead of requiring to specify enough digits in such ratios,
add the possibility to configure actual fractions like "1/3" (and
lists thereof, e.g. "1/3,2/3,0.9".

Alternatives would include tolerating slightly above as well, say
-10%/+1%, but that feels less exact and less safe (allowing even
slightly more could potentially overshoot available memory in some
cases).

If one of the newer CUSTOM_NUMASIZE_* host traits is required on a
(BigVM) flavor, then the host was already matched in placement, and
the filter can succeed early. Once we do not have any of the old-style
flavors anymore, we can thus remove this filter.

Co-Authored-By: Jakob Karge <jakob.karge@sap.com>
Change-Id: I45a45440591fefb17bed6839a9cd2495babd099d
Incorporated-Change-Id: I302ebea93dfe30eb72c9a0dfe42f5e8c956f228a
Incorporated-Change-Id: I3223aee433ba9009d2cd6387eeda8e70dbbb3cde