Func: bug 1849165: mig race with _populate_assigned_resources
Add a functional regression test for the referenced bug:

If a migration is initiated, and update_available_resource runs on the
destination between when the migration record is associated with the
destination and when the migration context is added to the instance, it
will raise a TypeError attempting to _populate_assigned_resources for
that instance, because that method attempts to access the
(as-yet-nonexistent) migration context.

Note that this doesn't fail the migration; it just leaves ugly logs. In
real life it probably also leaves other pieces of
update_available_resource unfinished on the destination.

Related-Bug: #1849165
Change-Id: I7e96cd24049c205f76a684a2e7425f85b4376f73