Require Migration object arg to migrate_instance_finish method
There is one place in the compute manager that is passing a dict
for the migration to the migrate_instance_finish method but a
Migration object should be passed. The dict works because the
Migration object uses the NovaObjectDictCompat mixin but that is
something we'd like to eventually remove since it's nice to know
if you're dealing with an object or a dict.

This converts the remaining uses of migrate_instance_finish that
were using a dict for the migration over to using an object.

Change-Id: Ice8b167d6fe026e7043a0899f399a62e25dcfcd1