Stop converting Migration objects to dicts for migrate_instance_start
The Migration object has the NovaObjectDictCompat mixin so we can
just pass the Migration object directly to the migrate_instance_start
method and it works like a dict so this change removes the unnecessary
obj_to_primitive conversion in a couple of spots. Apparently there were
no tests that actually cared about that conversion either.

Long-term the migrate_instance_start method should enfroce that an
object is passed rather than a dict but that's a much bigger effort
left for another day.

Change-Id: I1b7657b450676cd537b4afaf6773e940f65314a6