[stable-only] Disable cinder-backup service in nova-next job
We don't run any tests that require the cinder backup
service in the nova-next job nor do we care to so save
some time in stacking and space on the node, and avoid
spurious volume backup test failures, by not
installing that service.

Note that this was done in Train with commit
4188d08c20b67b4de834c60b41ec5987a9bd0136 but this is not
a cherry-pick because Ie3dc90862c895a8bd9bff4511a16254945f45478
which converted the nova-next job to zuulv3 format is not
in stable/stein.

Change-Id: I6991a231366908920847fa1b8375ece8d8b35268