doc: note the need to configure cinder auth in reclaim_instance_interval
Without configuring the [cinder] auth options in nova.conf, reclaiming
SOFT_DELETED instances will fail to cleanup attached volumes since there
is no user token in a periodic task and we have to rely on config to
use a service user for that cleanup with cinder. This adds a note to the
reclaim_instance_interval config option help to explain that.

Change-Id: I9da30b53b4d1bcaca7944700a67a19523f3bcf5d
Closes-Bug: #1849481