Reset instance to current vm_state if rolling back in resize_instance
You can resize a stopped instance and if the compute driver raises
InstanceFaultRollback from migrate_disk_and_power_off, the
_error_out_instance_on_exception decorator, used in the _resize_instance
method, will by default reset the instance vm_state to ACTIVE even though
the guest is stopped. The driver could raise InstanceFaultRollback if you
try resizing the root disk down on a non-volume-backed instance.

This builds on [1] and does the same thing as prep_resize [2] for
making sure the original vm_state is reset on InstanceFaultRollback.

[1] Ie4f9177f4d54cbc7dbcf58bd107fd5f24c60d8bb
[2] I17543ecb572934ecc7d0bbc7a4ad2f537fa499bc

Change-Id: Iff1f9f28a1e4ecf00368cbcac27b7687a5eb0dcf
Closes-Bug: #1551703