libvirt: flatten rbd image during cross-cell move spawn at dest
When migrating a non-volume-backed server across cells we will
create a temporary snapshot image and use that to spawn the
guest in the destination host/cell (like unshelve) and then
_finish_snapshot_based_resize_at_dest in the ComputeManager
will attempt to delete the snapshot image. For rbd images,
deleting the snapshot image will fail if we don't flatten it
first.

This builds on If3c9d1de3ce0fe394405bd1e1f0fa08ce2baeda8 and
adds logic to the libvirt driver such that, like the unshelve
scenario, if we are doing a cross-cell move we flatten the rbd
image so the ComputeManager can delete the temporary snapshot
image after the guest is spawned.

Part of blueprint cross-cell-resize

Change-Id: If56842da51688476072814e6e8309cf9e0a0210c