Cleanup libvirt test_mount unit tests
The primary goal of this change is to remove os.path.ismount mocks
like:

@mock.patch('os.path.ismount', side_effect=[False, True, True, True])

The 2 main problems with these mocks are 1) they're opaque and hard to
validate, and 2) they need to be rewritten if the code under test
changes the call order, which makes it harder to validate such
changes.

We replace these mocks with MountFixture, which simply correlates
calls to mount and unmount.

Change-Id: I3c51b99ca28093013b90da1a7360a6728eaf224d