vmware: image template cache expiration
While nova already cleaned up its original image-cache, our patches
also added images as templates into the mix that were not cleaned up
prior to this patch and thus filling up the datastores.

We sort by creation-time, so we can stop looping over tasks once we find
an event that's too old. All other VMs, we didn't find an event for,
yet, must be expired anyways.

Since images are kept in the owner's folder and not uploaded
individually for each project using them, we need to make sure we clean
up everywhere - even in projects without instances.

We want to make sure we only clean up image-cache templates and no other
VMs that might happen to lie in the "Images" folder. Therefore, we
validate the VM name conforms to the naming scheme we create image
template VMS with.

Since every nova-compute node will be running the image-cache cleanup,
we have to make sure this works in parallel. Therefore, we limit the
cleanup of image-cache templates to the datastores the nova-compute node
is responsible for - the ones configured in the datastore_regex config
option.

The logs during clean-up would show "Destroyed VM" messages, but one
could not find out, which kind of VM would be destroyed. Therefore, we
add an additional log message to point out that and which image-template
VM we're going to destroy.

Co-Authored-By: Johannes Kulik <johannes.kulik@sap.com>
Change-Id: I2fc631d6ce0a9339be9237d63ae7e86d94779dcc
Incorporated-Change-Id: I196bb9ab48867f314d9a5f3b7566384fa72778df
Incorporated-Change-Id: I9344a9bb8c0436cb5c87f3328fa3b9d31dedbdbc
Incorporated-Change-Id: I55b7fa7ebbe14b13f579ebc39dcae2549ddedc9a
Incorporated-Change-Id: I5fae822a08bcc06565c64f959553cf7082bb2423
Incorporated-Change-Id: I924fbc42f014ecf4c0342246d48f27b6ba1d5c77
Incorporated-Change-Id: I7429fca0175ec4593689466be6dcc0cb2482cb9f